Adaptacion a multi step form
Fixes #148
